pyproject.toml
[project]
name = "cli-agent-orchestrator"
version = "2.0.0"
description = "CLI Agent Orchestrator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # --- CAO core ---
    "fastapi>=0.104.0",
    "fastmcp>=2.14.0",
    "mcp>=1.23.0",
    "apscheduler>=3.10.4",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "libtmux>=0.51.0",
    "aiofiles>=24.1.0",
    "python-frontmatter>=1.1.0",
    "watchdog==6.0.0",
    # --- Merged from amplifier-amplifier-claude ---
    "aiohttp>=3.12.15",
    "anthropic>=0.69.0",
    "beautifulsoup4>=4.14.2",
    "claude-code-sdk>=0.0.20",
    "httpx>=0.28.1",
    "langchain>=0.2.1",
    "langchain-openai>=0.3.28",
    "markdownify>=1.2.0",
    "networkx>=3.4,<3.5",  # upper-pinned: networkx>=3.5 requires Python>=3.11, we target >=3.10
    "openai>=1.108.1",
    "pydantic-ai>=1.0.10",
    "pydantic-settings>=2.10.1",
    "pydot>=4.0.1",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.2",
    "rapidfuzz>=3.13.0",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
    "yt-dlp>=2025.9.26",
    # --- Shared (higher of the two lower bounds) ---
    "click>=8.2.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[project.scripts]
"cao" = "cli_agent_orchestrator.cli.main:cli"
"cao-server" = "cli_agent_orchestrator.api.main:main"
"cao-mcp-server" = "cli_agent_orchestrator.mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cli_agent_orchestrator"]

[tool.hatch.build]
artifacts = [
    "src/cli_agent_orchestrator/web_ui/**",
]

[dependency-groups]
dev = [
    # --- CAO dev ---
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest-xdist>=3.8.0",
    # --- Merged from amplifier-amplifier-claude ---
    "build>=1.2.2.post1",
    "debugpy>=1.8.14",
    "pyright>=1.1.407",
    "ruff>=0.11.10",
    "twine>=6.1.0",
    # --- Shared (higher of the two lower bounds) ---
    "pytest>=9.0.3",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.1.1",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = [
    "**/__pycache__",
    "**/.venv/**",
    "**/node_modules/**",
    ".git/**",
    ".ruff_cache/**",
    ".pytest_cache/**",
    "*.egg-info",
    "build/**",
    "dist/**",
    ".data/**",
    "ai_working/**",
]
typeCheckingMode = "basic"
reportMissingImports = false
pythonVersion = "3.10"

[tool.pytest.ini_options]
markers = [
    "asyncio: marks tests that use asyncio",
    "integration: marks integration tests",
    "e2e: marks end-to-end tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
asyncio_mode = "strict"
testpaths = ["test"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src --cov-report=term-missing -m 'not e2e'"
